+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added GetDirectoryListing of a specific path
 - Added stderr to output of ShellExecute functions
+- Added ShellCmdExecute function
 
 ### Fixed
 - Fixed XML warning, removed angle brackets in comment